MDEV 15532 Assertion `!log->same_pk' failed in row_log_table_apply_delete
The reason for the failure is that
thd->mdl_context.release_transactional_locks()
was called after commit & rollback even in cases where the current
transaction is still active.

Fixed by changing the return value for the following functions to
return -1 in the case the transation is still active:
- trans_begin()
- trans_commit()
- trans_commit_implicit()
- trans_rollback()
- trans_xa_commit()
- trans_xa_rollback()

In addition, trans_xa_commit() and trans_xa_rollback() will automatically
call release_transactional_locks() if the transaction is closed.
We can't do that for the other functions as the caller of many of these
are doing additional work (like close_thread_tables) before calling
release_transactional_locks().
trans_rollback_implict() didn't have to be changed as it doesn't interact
with XA and thus will always rollback the current transaction.

All callers of the above listed functions are updated to reflect the new
interface. The one exception is replication as there will not be any
active XA transaction when commit/rollback is called.

Other things:
- Added an assert in release_transactional_locks() to catch if anyone
  calls it with an active transaction.
- Injector::transaction::commit() now return error if transaction
  rolls back.  This was documented before but not done.
- Fixed missing abort_result_set() and wrong return calls in
  select_create::send_eof()